+++ pages/_app.js
@@ -4,6 +4,7 @@ import Head from "next/head";
 import Script from "next/script";
 import { useEffect, useState } from "react";
 import TagManager from "react-gtm-module";
+import Clarity from "@microsoft/clarity";
 import "styles/style.scss";
 
 const App = ({ Component, pageProps }) => {
@@ -28,6 +29,12 @@ const App = ({ Component, pageProps }) => {
         TagManager.initialize({ gtmId: config.params.tag_manager_id });
       }
     }, 5000);
+  }, []);
+
+  useEffect(() => {
+    const projectId = process.env.NEXT_PUBLIC_CLARITY_PROJECT_ID;
+    if (!projectId) return;
+    Clarity.init(projectId);
   }, []);
 
   return (
@@ -66,19 +73,6 @@ const App = ({ Component, pageProps }) => {
         src="//static.klaviyo.com/onsite/js/klaviyo.js?company_id=XpEmXv"
         strategy="afterInteractive"
       /> */}
-
-      {/* Microsoft Clarity */}
-      <Script
-        id="microsoft-clarity"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `(function(c,l,a,r,i,t,y){
-            c[a]=c[a]||function(){(c[a].q=c[a].q||[]).push(arguments)};
-            t=l.createElement(r);t.async=1;t.src="https://www.clarity.ms/tag/"+i;
-            y=l.getElementsByTagName(r)[0];y.parentNode.insertBefore(t,y);
-          })(window,document,"clarity","script","${process.env.NEXT_PUBLIC_CLARITY_PROJECT_ID}");`,
-        }}
-      />
 
       <Component {...pageProps} />
     </>